Change output signature dtype from float32 to int32
Fix dtype on generator Dataset.

+++ generative-proof-of-concept-CPU-preprocessing-in-memory.py
@@ -1390,7 +1390,7 @@ def objective(trial: optuna.Trial) -> float:
             dataset = tf.data.Dataset.from_generator(
                 lambda: generator,
                 output_signature=(
-                    tf.TensorSpec(shape=(MAX_SEQ_LENGTH,), dtype=tf.float32),
+                    tf.TensorSpec(shape=(MAX_SEQ_LENGTH,), dtype=tf.int32),
                     tf.TensorSpec(shape=(VOCABULARY_SIZE,), dtype=tf.float32)
                     # tf.TensorSpec(shape=(1, MAX_SEQ_LENGTH), dtype=tf.int32),       # Nested input
                     # tf.TensorSpec(shape=(1, VOCABULARY_SIZE), dtype=tf.float32)   # Nested one-hot label
